+++ packages/ui/src/index.ts
@@ -10,6 +10,7 @@ export { Text } from './components/Text/Text'
 export * as Dialog from './components/Dialog/Dialog'
 export { Badge, type BadgeProps } from './components/Badge/Badge'
 export { ConditionalWrapper } from './components/ConditionalWrapper'
+export * from './components/ThemeProvider'
 
 export * from './icons'
