+++ packages/ui/src/theme/index.ts
@@ -3,3 +3,4 @@ export * from './media-query'
 export * from './globalStyles'
 export * from './colors/colors'
 export * from './typography'
+export * from './margins'
